Weight based path finding & procedurally generate path
Made a script that gives each hex tile a weight, based on which it then generates the path. Some hexes are given a weight of 1, which is water. I also started to work on procedurally placing the tiles for the path. For that all that is left is to rotate correctly the placed tile.